Separate the structs VariableReference, ConstraintReference, and ConstraintTypeReference from the Model class.
